SignUp/SignIn using Google

+++ controllers/api/v1/post_api.js
@@ -1,5 +1,6 @@
 const Post = require('../../../models/post');
 const Comments = require('../../../models/comment');
+const User = require('../../../models/user');
 module.exports.index = async function(req,res){
     let posts = await Post.find({})
     .populate('user')
@@ -20,16 +21,23 @@ module.exports.index = async function(req,res){
 module.exports.destroy = async (req,res)=>{
     try {
         let post = await Post.findById(req.params.id);
-
+        if(post.user == req.user.id){
             post.remove();//removing whole collection in db(mongoose)
             await Comments.deleteMany({post:req.params.id});
-           
-        return res.status(200).json({
-            message:"Post deleted Succ"
-        })
-    } catch (error) {
-        return res.status(400).json({
-            message:"Post deleted unSucc"
+ 
+            return res.status(200).json({
+                message:"Post deleted Succ"
+            })
+
+        }else{
+            return res.status(401).json({
+                message:"UnAuthorized"
+            })
+        }
+    }catch(error) {
+        console.log("*****error",error);
+        return res.status(500).json({
+            message:"Internal Server Error"
         })
     }
 }
